updates
order service
product service

+++ irisMVC/datamodels/order.go
@@ -4,7 +4,7 @@ type Order struct {
 	ID int64 `sql:"ID"`
 	UserId int64 `sql:"userId"`
 	ProductId int64 `sql:"productId"`
-	OrderStatus int64 `sql:"orderStatus"`
+	OrderStatus int `sql:"orderStatus"`
 }
 
 const(
